Fixed token Expire time,Changes on SignalR
Fixed token Expire time,Changes on SignalR
